requirements.txt
selenium
pandas
openpyxl
Pillow
faker
numpy
requests
# tk is usually included in python

